Refactor: 콘텐츠, 주차별 정보 관련 컴포넌트/데이터 네이밍 통일
